Update CheckIfElementIsVisible method
Will use FindElements function. If ==1 >> true, else >> false.

Needed this when trying to assert if element is visible. The way to check this is to assert that there are no element on the page. If using Display() - it only check if the elements value of Display is true or fase, but if the element is not on the page - the test fails as it can not find the element at all to check.

Could use explicit wait for that with the ExpectedCondition "Element Exists" - but wait is cumbersome. Much easier to check with Count.